requirements.txt
PyQt<5.0
PyOpenGL
numpy
Cython
Pillow
